refactor(db): Extract txn_registry tests to separate file
Moves 951 lines of tests from txn_registry.rs into txn_registry_tests.rs
using #[path] attribute, reducing the implementation file by 69%.

Closes #331
